move table generation code into a separate file.
